# shared/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: FINN Licensing API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /validate:
    post:
      operationId: publicValidate
      tags: [public]
      summary: Validate a license
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateRequest"
      responses:
        "200":
          description: Validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidateResponse"
        "429":
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidateResponse"
  /products:
    get:
      operationId: publicProducts
      tags: [public]
      summary: List available products
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/PublicProduct"
                required:
                  - data
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /update-check:
    get:
      operationId: publicUpdateCheck
      tags: [public]
      summary: Check for updates
      parameters:
        - name: product_id
          in: query
          required: true
          schema:
            type: string
        - name: license
          in: query
          required: true
          schema:
            type: string
        - name: fingerprint
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Update info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateCheckResponse"
  /download:
    get:
      operationId: publicDownload
      tags: [public]
      summary: Download plugin
      parameters:
        - name: product_id
          in: query
          required: true
          schema:
            type: string
        - name: license
          in: query
          required: true
          schema:
            type: string
        - name: fingerprint
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Plugin ZIP file
          content:
            application/zip:
              schema:
                type: string
                format: binary
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    ValidateRequest:
      type: object
      properties:
        key:
          type: string
        fingerprint:
          type: string
      required:
        - key
        - fingerprint
    ValidateResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            valid:
              type: boolean
          required:
            - valid
      required:
        - data
    PublicProduct:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        version:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
      required:
        - id
        - name
        - slug
...
    UpdateCheckResponse:
      type: object
      properties:
        version:
          type: string
          nullable: true
        download_url:
          type: string
          nullable: true
        slug:
          type: string
          nullable: true
        tested:
          type: string
          nullable: true
        requires:
          type: string
          nullable: true
        requires_php:
          type: string
          nullable: true
        sections:
          type: object
          properties:
            changelog:
              type: string
          nullable: true